requirements.txt
openai>=1.0.0
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
python-dotenv>=1.0.0
puresnmp==2.0.1
loguru>=0.7.2
pytest>=7.4.0
httpx>=0.24.1
python-multipart>=0.0.6
rich>=13.9.0
